# server/config/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Mean Application API",
    "description": "Mean Application API",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "localhost:3000",
  "basePath": "/api/",
...
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/opportunity/create": {
      "post": {
        "tags": [
          "Opportunities"
        ],
        "description": "Adds an opportunity to the opportunities collection",
        "parameters": [
          {
            "name": "create",
            "in": "body",
            "description": "Create an opportunity document",
            "schema": {
              "type": "object",
              "properties": {
                "opportunity": {
                  "type": "object",
                  "$ref": "#/definitions/Opportunity"
                }
              }
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Opportunity is created and its data is returned",
            "schema": {
              "$ref": "#/definitions/Opportunity"
            }
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
    "/opportunity/get": {
      "post": {
        "tags": [
          "Opportunities"
        ],
        "description": "Get the opportunities in the collection, which the user sending the request has access to, with optional filters",
        "parameters": [
          {
            "name": "get",
            "in": "body",
            "description": "Finds opportunities with the sent filters",
            "schema": {
              "type": "object",
              "properties": {
                "filters": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string"
                    }
                  }
                },
                "onlyUserOpportunities": {
                  "type": "boolean"
                }
              }
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Returns the opportunities matching the filters and which user has access to in the collection",
            "schema": {
              "type": "object",
              "properties": {
                "opportunities": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/Opportunity"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
    "/demo/create": {
      "post": {
        "tags": [
          "Demos"
        ],
        "description": "Adds an demo to the demos collection",
        "parameters": [
          {
            "name": "create",
            "in": "body",
            "description": "Create a demo document",
            "schema": {
              "type": "object",
              "properties": {
                "demo": {
                  "type": "object",
                  "$ref": "#/definitions/Demo"
                }
              }
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Demo is created and its data is returned",
            "schema": {
              "$ref": "#/definitions/Demo"
            }
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
    "/demo/get": {
      "post": {
        "tags": [
          "Demos"
        ],
        "description": "Get the demos in the collection, which the user sending the request has access to, with optional filters",
        "parameters": [
          {
            "name": "get",
            "in": "body",
            "description": "Finds demos with the sent filters",
            "schema": {
              "type": "object",
              "properties": {
                "filters": {
                  "type": "object",
                  "properties": {
                    "opportunity": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Returns the demos matching the filters and which user has access to in the collection",
            "schema": {
              "type": "object",
              "properties": {
                "demos": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/Demo"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
...
  },
  "securityDefinitions": {
    "Authorization": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "definitions": {
...
    "Opportunity": {
      "required": [
        "name",
        "description",
        "companyClient"
      ],
      "properties": {
        "_id": {
          "type": "string",
          "uniqueItem": true,
          "readOnly": true
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "companyClient": {
          "type": "string"
        },
        "estimatedTotal": {
          "type": "number"
        },
        "automation": {
          "type": "boolean"
        },
        "docManager": {
          "type": "boolean"
        },
        "digitization": {
          "type": "boolean"
        },
        "hardware": {
          "type": "boolean"
        },
        "demo": {
          "type": "object",
          "$ref": "#/definitions/Demo"
        },
        "proposal": {
          "type": "object",
          "$ref": "#/definitions/Demo"
        },
        "createdAt": {
          "type": "string",
          "readOnly": true
        },
        "createdBy": {
          "type": "string",
          "readOnly": true
        },
        "modifieddAt": {
          "type": "string",
          "readOnly": true
        },
        "modifiedBy": {
          "type": "string",
          "readOnly": true
        }
      }
    },
    "Demo": {
      "required": [
        "dmeoType"
      ],
      "properties": {
        "_id": {
          "type": "string",
          "uniqueItem": true,
          "readOnly": true
        },
        "opportunity": {
          "type": "string",
          "readOnly": true
        },
        "dmeoType": {
          "type": "string"
        },
        "possibleDate_1": {
          "type": "string"
        },
        "possibleDate_2": {
          "type": "string"
        },
        "possibleDate_3": {
          "type": "string"
        },
        "selectedDate": {
          "type": "string"
        },
        "genericDemos": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "workflowFile": {
          "type": "object",
          "properties": {
            "data": {
              "type": "buffer"
            },
            "contentType": {
              "type": "string"
            }
          }
        },
        "createdAt": {
          "type": "string",
          "readOnly": true
        },
        "createdBy": {
          "type": "string",
          "readOnly": true
        },
        "modifieddAt": {
          "type": "string",
          "readOnly": true
        },
        "modifiedBy": {
          "type": "string",
          "readOnly": true
        }
      }
    }
  }
}